requirements.txt
torch
requests
praw
python-dotenv
scikit-learn
